chore: fix swiftlint error
